# 8-week/api/gen/openapi/gateway.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Auth API",
    "description": "RESTful gateway API for authentication and tokens.",
    "version": "1.0.0",
    "contact": {
      "name": "Daniil Kalts",
      "email": "gopher@gmail.com"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate user and return tokens",
        "description": "Authenticates a user with email and password credentials. Returns a new access/refresh token pair for authorized API access.",
        "operationId": "loginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "🔐 Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Invalidate the current session",
        "description": "Invalidates the session associated with the provided refresh token. Requires a valid refresh_token and returns an empty response on success.",
        "operationId": "logoutUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Logout\nThe access token should be passed via the Authorization header at the gateway level.\nThis RPC revokes the session identified by the provided refresh token.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LogoutRequest"
            }
          }
        ],
        "tags": [
          "🔐 Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "summary": "Rotate access and refresh tokens",
        "description": "Issues a new access/refresh token pair by rotating the provided refresh token. Requires a valid refresh_token and returns replacement tokens.",
        "operationId": "refreshTokens",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RefreshResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "Refresh\nToken rotation occurs: the provided refresh token is invalidated and a new pair is issued.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RefreshRequest"
            }
          }
        ],
        "tags": [
          "🔐 Auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "summary": "Register a new user account",
        "description": "Creates a new user account using name, email, and password credentials. Returns the created user profile and a fresh access/refresh token pair.",
        "operationId": "registerUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterRequest"
            }
          }
        ],
        "tags": [
          "🔐 Auth"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "Login"
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "tokens": {
          "$ref": "#/definitions/v1TokenPair"
        }
      }
    },
    "v1LogoutRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      },
      "description": "Logout\nThe access token should be passed via the Authorization header at the gateway level.\nThis RPC revokes the session identified by the provided refresh token."
    },
    "v1RefreshRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "type": "string"
        }
      },
      "description": "Refresh\nToken rotation occurs: the provided refresh token is invalidated and a new pair is issued."
    },
    "v1RefreshResponse": {
      "type": "object",
      "properties": {
        "tokens": {
          "$ref": "#/definitions/v1TokenPair"
        }
      },
      "description": "Token rotation occurs: the provided refresh token is invalidated and a new pair is issued."
    },
    "v1RegisterRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "Register"
    },
    "v1RegisterResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/v1RegisterResponseUser"
        },
        "tokens": {
          "$ref": "#/definitions/v1TokenPair"
        }
      }
    },
    "v1RegisterResponseUser": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
...
    "v1TokenPair": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "accessTokenExpiresIn": {
          "type": "string",
          "format": "int64"
        },
        "refreshTokenExpiresIn": {
          "type": "string",
          "format": "int64"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "JWT access token. Example: Bearer \u003ctoken\u003e",
      "name": "Authorization",
      "in": "header"
    }
  }
}